Split the two testImageScaling methods into six.
We had one test using the image as property, and one getting the image from data.
They both tested the tag and the scale methods.
I have now split those two.

And for the scale method we tried the 'mini' scale, which is the same size as the original image we use.
So use thumb instead. That should flush out some more problems.
And add tests calling scale with width=10.

I am getting an error in the combination of data and tag, and suspected data and scale would fail as well.
This would make that clearer.

Currently only `testImageScalingTagViaData` is failing.
In the tag we expect a width and height of 10, because we ask for this.
Instead we get 400 by 400.  Quite strange, since the original is 200 by 200...
Might need another fix in plone.namedfile or plone.scale.